refactor: activity route and component
