+++ src/app/(event)/layout.tsx
